requirements.txt
# Backend Server
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0

# LlamaIndex Core
llama-index>=0.9.11
llama-index-llms-openai>=0.1.5
llama-index-readers-file>=0.1.4
llama-parse>=0.1.0

# OpenAI
openai>=1.3.0

# Data Processing
pandas>=2.1.0
numpy>=1.25.2

# Financial APIs
yfinance>=0.2.31
requests>=2.31.0
beautifulsoup4>=4.12.3
tradingview-ta>=3.3.0

# Optional - Recommended
matplotlib>=3.7.0
scikit-learn>=1.0.2 